requirements.txt
Flask==3.1.0
groq==0.15.0
gunicorn==20.1.0
openai==1.60.2
python-dotenv==1.0.1